Refactor: 상수 파일 경로 변경
